// src_bak/utils/types.ts
export interface ISpell {
  id: string;
  name: string;
  strength: number;
  element: IElement;
  description: string;
}

export type IElement = {
  code: string;
  color: colorType;
  description: string;
  school: ISchool;
};

export type ISchool = {
  id: schoolType;
  description: string;
  name: string;
};

export type IPlayerSpell = ISpell & {
  copy: number;
  selected: boolean;
  owner: "hero" | "enemy";
  updates: IPlayerSpellUpdate[];
  created_at: Date;
};
export type IPlayerResource = IResource & {
  quantity: number;
  created_at: Date;
};
export type INPC = ICharacter & {
  image: string;
  dialogue: string | null;
  created_at: Date;
};
export type IHero = ICharacter & {
  element: IElement;
};
export type IPlayerHero = IHero & { selected: boolean; created_at: Date };

export type IEnemy = ICharacter & {
  element: IElement;
};

export type IEnemyFight = IEnemy & { spells: IPlayerSpell[] };

export type ISpellUpdate = {
  id: string;
  school: ISchool;
  mana: number;
  effect: spellEffectType;
  action: IUpdateAction;
  price: { action: spellUpdatePrice; strength: number | string } | null;
  name: string;
  description: string;
  resource_base: IResource[];
};

export type IPlayerSpellUpdate = ISpellUpdate & {
  created_at: Date;
  resource_base: IResource[];
};
// ...
export type IStoryAction = {
  id: string;
  type: storyChangeType;
  item: string;
  data?: string | null;
};

export type ICharacter = {
  id: string;
  name: string;
  description: string;
};
// ...
export type IStory = {
  type: storyType;
  id: string;
  nextStory: string;
  name: string;
  image: string;
  actions: IStoryAction[];
  open?: boolean;
};

export type IFight = {
  id: string;
  name: string;
  hero_num: number;
  enemy: IEnemy;
  hero_elements: colorType[];
};
// ...
export type IResource = {
  id: string;
  name: string;
  school: ISchool;
  description: string;
  commonality: number;
};

export type storyType = "dialogue" | "fight" | "reel";

export type adventureType =
  | "character"
  | "story"
  | "arena"
  | "torunament"
  | "event";
export type screenState = "start" | "opening" | "main" | "dialogue";
export type enemyExpLevel =
  | "novice"
  | "apprentice"
  | "practitioner"
  | "master"
  | "grandmaster";
export type colorType =
  | "violet"
  | "grey"
  | "red"
  | "indigo"
  | "green"
  | "blue"
  | "white"
  | "orange"
  | "yellow"
  | "rose"
  | "teal"
  | "sand"
  | "lime"
  | "black"
  | "brown";
export type storyChangeType =
  | "addNpc"
  | "setAdventure"
  | "openStory"
  | "addHero"
  | "addSpells"
  | "addUpdate";

export type spellEffectType =
  | "h_heal"
  | "h_trumpremove"
  | "h_trumpset"
  | "h_redraw"
  | "h_mana"
  | "e_redraw"
  | "e_drop"
  | "e_reduice"
  | "h_enforce";

export type herosSelectionError = "none" | "less" | "more" | "incorrect" | null;